+++ src/llm-providers/_common/register/$provideLlmToolsForWizardOrCli.ts
@@ -1,7 +1,6 @@
 import { join } from 'path';
 import { Promisable } from 'type-fest';
-import { DEFAULT_EXECUTION_CACHE_DIRNAME } from '../../../config';
-import { DEFAULT_REMOTE_SERVER_URL } from '../../../config';
+import { DEFAULT_EXECUTION_CACHE_DIRNAME, DEFAULT_REMOTE_SERVER_URL } from '../../../config';
 import { EnvironmentMismatchError } from '../../../errors/EnvironmentMismatchError';
 import { UnexpectedError } from '../../../errors/UnexpectedError';
 import type { LlmExecutionTools } from '../../../execution/LlmExecutionTools';
@@ -11,9 +10,7 @@ import { promptbookTokenToIdentification } from '../../../remote-server/socket-t
 import { $provideFilesystemForNode } from '../../../scrapers/_common/register/$provideFilesystemForNode';
 import { $EnvStorage } from '../../../storage/env-storage/$EnvStorage';
 import { FileCacheStorage } from '../../../storage/file-cache-storage/FileCacheStorage';
-import type { string_app_id } from '../../../types/typeAliases';
-import type { string_promptbook_token } from '../../../types/typeAliases';
-import type { string_url } from '../../../types/typeAliases';
+import type { string_app_id, string_promptbook_token, string_url } from '../../../types/typeAliases';
 import { $isRunningInNode } from '../../../utils/environment/$isRunningInNode';
 import type { really_any } from '../../../utils/organization/really_any';
 import { RemoteLlmExecutionTools } from '../../remote/RemoteLlmExecutionTools';
@@ -120,7 +117,9 @@ export async function $provideLlmToolsForWizardOrCli(
             identification,
         });
     } else if (strategy === 'BRING_YOUR_OWN_KEYS') {
-        llmExecutionTools = await $provideLlmToolsFromEnv();
+        llmExecutionTools = await $provideLlmToolsFromEnv({
+            title: 'LLM Tools for wizard or CLI with BYOK strategy',
+        });
     } else {
         throw new UnexpectedError(`\`$provideLlmToolsForWizardOrCli\` wrong strategy "${strategy}"`);
     }